Move each question to a component #5
This changes makes the code easier to read and the header allows me to add styling later by making changes only in one place
