+++ src/components/FAQ/Phones.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import Question from './Question';
+
+export default function Phones() {
+  return (
+    <>
+      <Question>
+        Can I use my phone to take pictures during the ceremony?
+      </Question>
+      <p className="body1">
+        No, we have chosen to have an unplugged ceremony and have hired professionals to photograph,
+        record, and live stream our ceremony. Please respect our wishes by turning off your phone.
+      </p>
+    </>
+  );
+}
