+++ src/components/FAQ/index.js
@@ -1,0 +1,3 @@
+import FAQ from './FAQ';
+
+export default FAQ;
